Cible web : import conditionnel de la base et entiers 64 bits portables
Deux blocages distincts, tous deux révélés par le build web de la CI (§3.9) —
et c'est exactement pourquoi §6.2 demande de s'occuper du web tôt.

`drift/native.dart` s'appuie sur `dart:ffi` : l'importer sans condition casse
la **compilation** web, pas seulement l'exécution. L'ouverture de la base
passe donc par un import conditionnel. La variante web renvoie `null` :
`sqlite3.wasm` + IndexedDB (§3.2) n'est pas branché, et l'app le reconnaît en
affichant des vues vides plutôt qu'en prétendant persister quelque chose.

JavaScript n'a pas d'entier 64 bits : `PlatformInt64` vaut `int` en natif et
`BigInt` en web. Les horodatages et identifiants traversent maintenant le pont
via `PlatformInt64Util.from` à l'entrée et `.toInt()` à la sortie, ce qui
compile pour les deux cibles sans rien perdre.

`flutter build web` passe en local. Les tests natifs restent verts.

+++ app/lib/features/stats/stats_view.dart
@@ -78,7 +78,7 @@ class StatsView extends StatelessWidget {
             title: l10n.statsTimePerPlace,
             child: _BucketList(
               buckets: stats.places,
-              valueOf: (bucket) => formatDuration(bucket.durationMs),
+              valueOf: (bucket) => formatDuration(bucket.durationMs.toInt()),
             ),
           ),
           _Section(
